move input methods, add docstring to inputmethods
